Moved mixins to a directory
