+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/definitions/OAS2Parser.java
@@ -74,7 +74,8 @@ import org.wso2.carbon.apimgt.api.model.CORSConfiguration;
 import org.wso2.carbon.apimgt.api.model.Scope;
 import org.wso2.carbon.apimgt.api.model.SwaggerData;
 import org.wso2.carbon.apimgt.api.model.URITemplate;
-import org.wso2.carbon.apimgt.impl.definitions.APIParserConstants;
+import org.wso2.carbon.apimgt.impl.definitions.mixin.IgnoreOriginalRefMixin;
+import org.wso2.carbon.apimgt.impl.definitions.mixin.ResponseSchemaMixin;
 
 import java.io.IOException;
 import java.util.ArrayList;
